Rename eventstoredb -> kurrentdb
