+++ kurrentdb/eventstore/errors.go
