+++ kurrentdb/eventstore/options.go
