+++ kurrentdb/eventstore/stream_iterator.go
